Move UI render coverage out of test package
