--- packages/test/src/test/framework.test.browser.tsx
@@ -1,173 +1,0 @@
-import decamelize from 'decamelize'
-import cx from 'clsx'
-import * as assert from '@remix-run/assert'
-import { describe, it } from '@remix-run/test'
-import { on, type Handle } from '@remix-run/ui'
-import { render } from '@remix-run/ui/test'
-
-describe('Counter', () => {
-  function Counter(handle: Handle<{ count?: number }>) {
-    let count = handle.props.count ?? 0
-    return () => (
-      <div>
-        <h3>Counter</h3>
-        <div>
-          <button
-            data-action="decrement"
-            mix={[
-              on('click', () => {
-                count--
-                handle.update()
-              }),
-            ]}
-          >
-            -
-          </button>
-          <span
-            data-testid="count"
-            style={{ fontSize: '24px', minWidth: '2ch', textAlign: 'center' }}
-          >
-            {count}
-          </span>
-          <button
-            data-action="increment"
-            mix={[
-              on('click', () => {
-                count++
-                handle.update()
-              }),
-            ]}
-          >
-            +
-          </button>
-        </div>
-      </div>
-    )
-  }
-
-  it('renders with initial count of 0 when not specified', (t) => {
-    let { $, cleanup } = render(<Counter />)
-    t.after(cleanup)
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 0)
-  })
-
-  it('renders with a provided initial count', (t) => {
-    let { $, cleanup } = render(<Counter count={5} />)
-    t.after(cleanup)
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 5)
-  })
-
-  it('increments the count', async (t) => {
-    let { $, act, cleanup } = render(<Counter />)
-    t.after(cleanup)
-    await act(() => $('[data-action="increment"]')?.click())
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 1)
-    await act(() => $('[data-action="increment"]')?.click())
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 2)
-    await act(() => $('[data-action="increment"]')?.click())
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 3)
-  })
-
-  it('decrements the count', async (t) => {
-    let { $, act, cleanup } = render(<Counter count={3} />)
-    t.after(cleanup)
-    await act(() => $('[data-action="decrement"]')?.click())
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 2)
-    await act(() => $('[data-action="decrement"]')?.click())
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 1)
-    await act(() => $('[data-action="decrement"]')?.click())
-    assert.equal(Number($('[data-testid="count"]')!.textContent), 0)
-  })
-})
-
-describe('FieldLabel (using decamelize)', () => {
-  // Demonstrates that ESM third-party libraries are importable from test modules
-  function FieldLabel(handle: Handle<{ name: string }>) {
-    return () => (
-      <span data-testid="label">{decamelize(handle.props.name, { separator: ' ' })}</span>
-    )
-  }
-
-  it('renders a single word unchanged', (t) => {
-    let { $, cleanup } = render(<FieldLabel name="name" />)
-    t.after(cleanup)
-    assert.equal($('[data-testid="label"]')?.textContent, 'name')
-  })
-
-  it('converts camelCase to spaced words', (t) => {
-    let { $, cleanup } = render(<FieldLabel name="firstName" />)
-    t.after(cleanup)
-    assert.equal($('[data-testid="label"]')?.textContent, 'first name')
-  })
-
-  it('handles multiple humps', (t) => {
-    let { $, cleanup } = render(<FieldLabel name="dateOfBirth" />)
-    t.after(cleanup)
-    assert.equal($('[data-testid="label"]')?.textContent, 'date of birth')
-  })
-})
-
-describe('MobileMenu (using clsx)', () => {
-  function MobileMenu(handle: Handle<{ isOpen: boolean }>) {
-    return () => {
-      let { isOpen } = handle.props
-
-      return (
-        <nav
-          aria-label="Mobile navigation"
-          className={cx('mobile-menu', {
-            'mobile-menu--open': isOpen,
-            'mobile-menu--closed': !isOpen,
-          })}
-        >
-          <a href="/docs">Docs</a>
-          <a href="/blog">Blog</a>
-        </nav>
-      )
-    }
-  }
-
-  it('resolves browser-oriented package exports for default imports', (t) => {
-    let { $, cleanup } = render(<MobileMenu isOpen />)
-    t.after(cleanup)
-
-    assert.equal($('[aria-label="Mobile navigation"]')?.className, 'mobile-menu mobile-menu--open')
-  })
-})
-
-describe('DOM Tests', () => {
-  it('can interact with DOM', async () => {
-    let div = document.createElement('div')
-    div.textContent = 'Hello'
-    assert.equal(div.textContent, 'Hello')
-  })
-
-  it('can test fetch API', async () => {
-    let response = await fetch('data:text/plain,hello')
-    let text = await response.text()
-    assert.equal(text, 'hello')
-  })
-
-  it.skip('skip: can skip tests', () => {
-    assert.equal(true, false)
-  })
-
-  it.todo('todo: can mark tests as todo')
-})
-
-describe('render/cleanup', () => {
-  it('cleanup removes the container from the DOM', () => {
-    let { container, cleanup } = render(<div data-testid="manual">hello</div>)
-    assert.equal(document.body.contains(container), true)
-    cleanup()
-    assert.equal(document.body.contains(container), false)
-  })
-})
-
-describe.skip('skip: Skipped Test Suite', () => {
-  it('would fail', () => {
-    assert.equal(true, false)
-  })
-})
-
-describe.todo('todo: Test Suite')
