+++ packages/test/tsconfig.json
@@ -11,9 +11,7 @@
     "verbatimModuleSyntax": true,
     "erasableSyntaxOnly": true,
     "isolatedModules": true,
-    "skipLibCheck": true,
-    "jsx": "react-jsx",
-    "jsxImportSource": "@remix-run/ui"
+    "skipLibCheck": true
   },
   "exclude": ["dist"]
 }
